requirements.txt
streamlit
pandas
matplotlib.pyplot
numpy
japanize_matplotlib
plotly.express